Moved ToggleDrawerButton to /components instead of /screens/shared
